\name{rtc-package}
\alias{rtc-package}
\alias{rtc}
\docType{package}
\title{Tree Clustering}
\description{
RTC performs cluster analysis using trees. A tree approximates
the probability density function from which data has been drawn.
A set of such trees is inferred using a Metropolis-Hastings sampler,
and returned as a partitioning of the parameter space into rectangular segments.
}
\details{
\tabular{ll}{
Package: \tab rtc\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-08-14\cr
License: \tab Proprietary\cr
}
RTC provides a single function \code{tree.clustering}, which performs
clustering of a given dataset.
}
\author{Peter Kuma <pkuma@pixelfederation.com>}
\references{}
\keyword{package}
